requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=pypi_0
aiohttp=3.7.4.post0=pypi_0
albumentations=1.1.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
appdirs=1.4.4=pypi_0
argcomplete=1.12.3=pypi_0
async-timeout=3.0.1=pypi_0
attrs=21.2.0=pypi_0
backcall=0.2.0=pypi_0
black=21.4b2=pypi_0
blas=1.0=mkl
boto=2.49.0=pypi_0
bottleneck=1.3.2=py38heb32a55_1
brotlipy=0.7.0=py38h497a2fe_1001
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.2=pypi_0
catalogue=2.0.4=py38h578d9bd_0
certifi=2021.10.8=py38h06a4308_0
cffi=1.14.6=py38ha65f79e_0
chardet=4.0.0=py38h578d9bd_1
charset-normalizer=2.0.4=pypi_0
click=7.1.2=pyh9f0ad1d_0
clip=1.0=pypi_0
cloudpickle=1.6.0=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
crcmod=1.7=pypi_0
cryptography=3.4.7=py38ha5dfef3_0
cudatoolkit=11.0.221=h6bb024c_0
cupy=9.3.0=py38h9612f98_0
cycler=0.10.0=pypi_0
cymem=2.0.5=py38h709712a_2
cython=0.29.24=pypi_0
cython-blis=0.7.4=py38h5c078b8_0
dataclasses=0.8=pyhc8e2a94_3
decorator=5.0.9=pypi_0
detectron2=0.5+cu110=pypi_0
en-core-web-sm=3.1.0=pypi_0
fasteners=0.16.3=pypi_0
fastrlock=0.6=py38h709712a_1
filelock=3.0.12=pypi_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.8.1=pypi_0
ftfy=6.0.3=pypi_0
future=0.18.2=pypi_0
fvcore=0.1.5.post20210812=pypi_0
gcs-oauth2-boto-plugin=2.7=pypi_0
google-apitools=0.5.32=pypi_0
google-auth=1.34.0=pypi_0
google-auth-oauthlib=0.4.5=pypi_0
google-reauth=0.1.1=pypi_0
grpcio=1.39.0=pypi_0
gsutil=4.68=pypi_0
gym=0.20.0=pypi_0
httplib2=0.19.1=pypi_0
huggingface-hub=0.0.17=pypi_0
hydra-core=1.1.0=pypi_0
idna=3.2=pypi_0
imageio=2.9.0=pypi_0
importlib-resources=5.2.2=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
iopath=0.1.8=pypi_0
ipython=7.26.0=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.18.0=pypi_0
jinja2=3.0.1=pyhd8ed1ab_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
learn2learn=0.1.6=pypi_0
libblas=3.9.0=11_linux64_mkl
libcblas=3.9.0=11_linux64_mkl
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=11.2.0=h69a702a_8
libgfortran5=11.2.0=h5c6108e_8
libgomp=9.3.0=h5101ec6_17
liblapack=3.9.0=11_linux64_mkl
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=pypi_0
markupsafe=2.0.1=py38h497a2fe_0
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.2=pypi_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
monotonic=1.6=pypi_0
multidict=5.1.0=pypi_0
murmurhash=1.0.5=py38h709712a_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.2=he6710b0_1
networkx=2.6.3=pypi_0
ninja=1.10.2=hff7bd54_1
numexpr=2.7.3=py38h22e1b3c_1
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
oauth2client=4.1.3=pypi_0
oauthlib=3.1.1=pypi_0
olefile=0.46=py_0
omegaconf=2.1.0=pypi_0
opencv-python=4.5.3.56=pypi_0
opencv-python-headless=4.5.3.56=pypi_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.4=py38h8c16a72_0
parso=0.8.2=pypi_0
pathspec=0.9.0=pypi_0
pathy=0.6.0=pyhd8ed1ab_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.1=py38h2c7a002_0
pip=21.2.2=py38h06a4308_0
portalocker=2.3.0=pypi_0
preshed=3.0.5=py38h709712a_1
prompt-toolkit=3.0.19=pypi_0
protobuf=3.17.3=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.2=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pydantic=1.8.2=py38h497a2fe_0
pydeprecate=0.3.1=pypi_0
pydot=1.4.2=pypi_0
pygments=2.10.0=pypi_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py38h578d9bd_3
python=3.8.11=h12debd9_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
pytorch-lightning=1.4.7=pypi_0
pytz=2021.3=pyhd3eb1b0_0
pyu2f=0.1.5=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
qpth=0.0.15=pypi_0
qudida=0.0.4=pypi_0
readline=8.1=h27cfd23_0
regex=2021.8.3=pypi_0
requests=2.26.0=pyhd8ed1ab_0
requests-oauthlib=1.3.0=pypi_0
retry-decorator=1.1.1=pypi_0
rsa=4.7.2=pypi_0
sacremoses=0.0.45=pypi_0
scikit-image=0.18.3=pypi_0
scikit-learn=0.24.2=py38hdc147b9_0
scipy=1.7.1=pypi_0
seaborn=0.11.2=pypi_0
setuptools=52.0.0=py38h06a4308_0
shellingham=1.4.0=pyh44b312d_0
six=1.16.0=pyhd3eb1b0_0
smart_open=5.1.0=pyhd8ed1ab_1
spacy=3.1.1=py38hfc89cab_0
spacy-legacy=3.0.8=pyhd8ed1ab_0
sqlite=3.36.0=hc218d9a_0
srsly=2.4.1=py38h709712a_0
tabulate=0.8.9=pypi_0
tensorboard=2.6.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
termcolor=1.1.0=pypi_0
thinc=8.0.8=py38hfc89cab_0
threadpoolctl=2.2.0=pyh8a188c0_0
tifffile=2021.8.30=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.3=pypi_0
toml=0.10.2=pypi_0
torchmetrics=0.5.1=pypi_0
torchvision=0.8.2=py38_cu110
tqdm=4.62.0=pyhd8ed1ab_0
traitlets=5.0.5=pypi_0
transformers=4.10.2=pypi_0
typer=0.3.2=pyhd8ed1ab_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd8ed1ab_0
wasabi=0.8.2=pyh44b312d_0
wcwidth=0.2.5=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.37.0=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
yarl=1.6.3=pypi_0
zipp=3.5.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
